# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.AuthLogoutResponse:
    properties:
      message:
        example: Logged out successfully
        type: string
    type: object
  auth.AuthRefreshResponse:
    properties:
      accessToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expiresInSeconds:
        example: 3600
        type: integer
      profile:
        $ref: '#/definitions/auth.UserProfile'
      scope:
        example: user:email read:user
        type: string
      tokenType:
        example: bearer
        type: string
      valid:
        example: true
        type: boolean
    type: object
  auth.UserProfile:
    properties:
      avatarUrl:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      username:
        type: string
      uuid:
        description: User UUID from database table
        type: string
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: This is the backend API for the Developer Portal, providing endpoints
    for managing organizations, teams, projects, components, and deployments.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Developer Portal Backend API
  version: "1.0"
paths:
...
  /api/auth/{provider}/handler/frame:
    get:
      consumes:
      - application/json
      description: Handle OAuth callback from provider and return authentication result
        in HTML frame
      parameters:
      - description: OAuth provider (githubtools or githubwdf)
        in: path
        name: provider
        required: true
        type: string
      - description: OAuth authorization code from provider
        in: query
        name: code
        required: true
        type: string
      - description: OAuth state parameter for security
        in: query
        name: state
        required: true
        type: string
      - description: Environment (development, staging, production)
        in: query
        name: env
        type: string
      - description: OAuth error parameter from provider
        in: query
        name: error
        type: string
      - description: OAuth error description from provider
        in: query
        name: error_description
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML page that posts authentication result to opener window
          schema:
            type: string
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties: true
            type: object
      summary: Handle OAuth callback
      tags:
      - authentication
  /api/auth/{provider}/start:
    get:
      consumes:
      - application/json
      description: Initiate OAuth authentication flow with the specified provider
      parameters:
      - description: OAuth provider (githubtools or githubwdf)
        in: path
        name: provider
        required: true
        type: string
      - description: Environment (development, staging, production)
        in: query
        name: env
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to OAuth provider authorization URL
          schema:
            type: string
        "400":
          description: Invalid provider or request parameters
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to generate authorization URL
          schema:
            additionalProperties: true
            type: object
      summary: Start OAuth authentication
      tags:
      - authentication
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user and invalidate authentication session
      parameters:
      - description: Environment (development, staging, production)
        in: query
        name: env
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged out
          schema:
            $ref: '#/definitions/auth.AuthLogoutResponse'
        "400":
          description: Invalid provider
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Logout failed
          schema:
            additionalProperties: true
            type: object
      summary: Logout user
      tags:
      - authentication
  /api/auth/refresh:
    get:
      consumes:
      - application/json
      description: Refresh or validate authentication token using Authorization header
        or session cookies
      parameters:
      - description: Environment (development, staging, production)
        in: query
        name: env
        type: string
      - description: Bearer token for validation
        example: '"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully refreshed token
          schema:
            $ref: '#/definitions/auth.AuthRefreshResponse'
        "400":
          description: Invalid provider
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication required or token invalid
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Token refresh failed
          schema:
            additionalProperties: true
            type: object
      summary: Refresh authentication token
      tags:
      - authentication
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"